Added a bit more functionality and cleaner code.
The website isn't much yet and to fix the funny loading / reloading delay per card I will have to make it so each card runs it's own update.
We'll see how that goes in the future.

For now this is just a simple project where I tested querying the web.

Enjoy!

PS. I haven't tested responsiveness on mobile devices.